refactor: move things to stay closer to theirs loved ones
