+++ .huskyrc
@@ -1,5 +1,5 @@
 {
   "hooks": {
-    "pre-commit": "echo lint-staged"
+    "pre-commit": "lint-staged"
   }
 }
