+++ package.json
@@ -14,8 +14,12 @@
     "codecov": "3.1.0",
     "eslint": "5.10.0",
     "eslint-config-prettier": "3.3.0",
+    "eslint-config-react-app": "3.0.6",
     "eslint-plugin-flowtype": "3.2.0",
+    "eslint-plugin-import": "2.14.0",
+    "eslint-plugin-jsx-a11y": "6.1.2",
     "eslint-plugin-prettier": "3.0.0",
+    "eslint-plugin-react": "7.11.1",
     "flow-bin": "0.67.1",
     "husky": "1.2.1",
     "lint-staged": "8.1.0",
