+++ web/.eslintrc.json
@@ -1,3 +1,4 @@
 {
-  "extends": "react-app"
+  "extends": ["react-app", "plugin:jsx-a11y/recommended"],
+  "plugins": ["jsx-a11y"]
 }
