+++ web/src/presentation/containers/__tests__/App.test.js
@@ -1,12 +1,11 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
-import { initializeDomainLayer } from '../../domain';
-import { App } from '../containers/App';
-import { initializeInternalization } from '../internalization';
+import { initializeDomainLayer } from '../../../domain';
+import { App } from '../App';
+import { initializeInternalization } from '../../internalization';
 import {Provider} from "react-redux";
 
-
-jest.mock('../../data/firebase/authentication', () => ({
+jest.mock('../../../data/firebase/authentication', () => ({
   onAuthStateChanged: jest.fn(() => () => {}),
 }));
 const { store } = initializeDomainLayer();
